+++ HISTORY.md
@@ -1,8 +1,9 @@
 1.0.0 / 2016-09-xx
 ==================
 
-* Completely rewritten in ES6, retaining compatiblity with last Node.js LTS (which is `4.5.0` at the moment of writing)
-* Switched to ioredis as Redis client module
+* Completely rewritten module core in ES6 while retaining compatiblity with last Node.js LTS (which is `4.5.0` at the moment of writing)
+* Switched to `ioredis` as Redis client module
+* Switched to `AVA` for tests
 * Using SEMVER
 
 0.2.0 / 2014-04-29